+++ js-web/react-admin/src/pages/Login.tsx
@@ -1,5 +1,5 @@
 import React, { SyntheticEvent, useState } from 'react'
-import { Redirect } from 'react-router-dom'
+import { Navigate } from 'react-router-dom'
 import axios from 'axios'
 
 import '../Login.css'
@@ -19,7 +19,7 @@ const Login = () => {
   }
 
   return (
-    redirect ? <Redirect to={'/'} /> :
+    redirect ? <Navigate to={'/'} /> :
     <main className="form-signin">
       <form onSubmit={submit}>
         <h1 className="h3 mb-3 fw-normal">Please Sign In</h1>