+++ js-web/react-admin/src/pages/products/Products.tsx
@@ -55,7 +55,7 @@ const Products = () => {
                   <td>{p.price}</td>
                   <td>
                     <div className="btn-group mr-2">
-                      <Link to={`/products/${p.id}/edit`}href="#" className="btn btn-sm btn-outline-secondary">Edit</Link>
+                      <Link to={`/products/${p.id}/edit`} className="btn btn-sm btn-outline-secondary">Edit</Link>
                       <a href="#" className="btn btn-sm btn-outline-secondary" onClick={() => del(p.id)}>Delete</a>
                     </div>
                   </td>