+++ js-web/react-admin/src/pages/products/ProductsEdit.tsx
@@ -2,7 +2,7 @@ import React, { SyntheticEvent, useEffect, useRef, useState } from 'react'
 import axios from 'axios'
 
 import Wrapper from '../../components/Wrapper'
-import { Redirect } from 'react-router-dom'
+import { Navigate } from 'react-router-dom'
 // import { Product } from '../../models/product'
 import ImageUpload from '../../components/ImageUpload'
 
@@ -41,7 +41,7 @@ const ProductsEdit = (props: any) => {
     setRedirect(true)
   }
 
-  if (redirect) return <Redirect to="/products" />
+  if (redirect) return <Navigate to="/products" />
   return (
     <Wrapper>
       <form onSubmit={submit}>